Refactored the slide menu so that it simply takes an adapter to provide the list items.

+++ libslidemenu/src/com/coboltforge/slidemenu/SlideMenuInterface.java
@@ -24,6 +24,6 @@ package com.coboltforge.slidemenu;
 public interface SlideMenuInterface {
 	
 	interface OnSlideMenuItemClickListener {
-		public void onSlideMenuItemClick(int itemId);
+		public void onSlideMenuItemClick(long itemId);
 	}
 }
